Warrior Class created

+++ app/src/main/java/jp/ac/uryukyu/ie/e235726/Enemy.java
@@ -34,7 +34,7 @@ public class Enemy extends LivingThing{
      * @param damage 受けたダメージ
      */
     public void wounded(int damage){
-        int hitPoint = getHitPoint();
+        double hitPoint = getHitPoint();
         setHitPoint(hitPoint - damage);
         if( hitPoint < 0 ) {
             setDead(true);
